+++ ft_memcpy.c
@@ -6,13 +6,15 @@
 /*   By: letsaguiar <letsaguiar@yandex.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 08:47:00 by letsaguiar        #+#    #+#             */
-/*   Updated: 2024/04/16 08:50:12 by letsaguiar       ###   ########.fr       */
+/*   Updated: 2024/04/17 10:22:29 by letsaguiar       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void    *ft_memcpy(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+
+void    *ft_memcpy(void *dest, const void *src, size_t n)
 {
-    unsigned int    i;
+    size_t  i;
 
     i = 0;
     while (i < n)
