+++ ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: letsaguiar <letsaguiar@yandex.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/16 10:02:20 by letsaguiar        #+#    #+#             */
-/*   Updated: 2024/04/16 10:05:39 by letsaguiar       ###   ########.fr       */
+/*   Updated: 2024/04/17 10:17:48 by letsaguiar       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned int    ft_strlcpy(char *dest, const char *src, unsigned int n)
+size_t  ft_strlcpy(char *dest, const char *src, size_t n)
 {
-    unsigned int    src_len;
+    size_t  src_len;
 
     src_len = ft_strlen(src);
     if (src_len + 1 < n)
